Moving from Oracle to HashiCorp
My main focus is now the hashistack. Will probably have to refine and focus for specific areas and use cases of the stack as I move on, but I will sgtart by having a general understanding and foundational skills on each of the components.

I keep a close eye and bandwidth on OCI and the related modules.